backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
python-dotenv
pypdf
